Resolved issue

+++ PDAEstimator_Domain/Entities/BerthDetails.cs
@@ -25,9 +25,9 @@ namespace PDAEstimator_Domain.Entities
         [DisplayName("Terminal")]
         public int? TerminalID { get; set; }
         public bool IsDeleted { get; set; }
-        public int? MaxLoa { get; set; }
-        public int? MaxBeam { get; set; }
-        public int? MaxArrivalDraft { get; set; }
+        public decimal? MaxLoa { get; set; }
+        public decimal? MaxBeam { get; set; }
+        public decimal? MaxArrivalDraft { get; set; }
         public string? TerminalName { get; set; }
 
     }
